+++ authentik/flows/views/interface.py
@@ -12,18 +12,15 @@ from authentik.flows.models import Flow
 # fail to parse it and render an empty page, so they are sent to the simplified
 # flow executor instead.
 MIN_WEBKIT_VERSION = (16, 4)
+MIN_EDGE_VERSION = (19,)
 
 
 def version_below(version: dict[str, Any], minimum: tuple[int, int]) -> bool:
-    """Check whether a parsed major/minor version is below minimum.
-
-    Returns False when the version is missing or not numeric, so an unrecognised
-    user agent keeps the default flow executor.
-    """
+    """Check whether a parsed major/minor version is below minimum."""
     try:
         major = int(version["major"])
         minor = int(version["minor"] or 0)
-    except (KeyError, TypeError, ValueError):
+    except KeyError, TypeError, ValueError:
         return False
     return (major, minor) < minimum
 
@@ -45,10 +42,9 @@ class FlowInterfaceView(InterfaceView):
             return True
         # Only use SFE for Edge 18 and older, after Edge 18 MS switched to chromium which supports
         # the default flow executor
-        if (
-            ua["user_agent"]["family"] == "Edge"
-            and int(ua["user_agent"]["major"]) <= 18  # noqa: PLR2004
-        ):  # noqa: PLR2004
+        if ua["user_agent"]["family"] == "Edge" and version_below(
+            ua["user_agent"], MIN_EDGE_VERSION
+        ):
             return True
         # https://github.com/AzureAD/microsoft-authentication-library-for-objc
         # Used by Microsoft Teams/Office on macOS, and also uses a very outdated browser engine
